:construction:Primeira versão de testes :bug: Corrigir a adição do item na lista de pedidos, a quantidade do item pedido pode ser fracionada. :bug: Corrigir o nome do progress dialog de Importacao/Exportacao de dados.

+++ app/src/main/java/com/br/minasfrango/data/model/ItemPedido.java
@@ -21,7 +21,7 @@ public class ItemPedido implements Serializable {
 
     private long id;
 
-    private int quantidade;
+    private double quantidade;
 
     private double valorTotal;
